ai-kanban-fastapi-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
openai>=1.40.0
